requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
boto3==1.42.3
botocore==1.42.3
click==8.3.1
coloredlogs==15.0.1
fastapi==0.123.8
flatbuffers==25.9.23
h11==0.16.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.16.3
packaging==25.0
pillow==12.0.0
protobuf==6.33.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
s3transfer==0.16.0
six==1.17.0
starlette==0.50.0
sympy==1.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
